requirements.txt
djangorestframework==3.12.1
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.34
django-js-asset==1.2.2
typing-extensions==3.7.4.3
pandas==1.1.4
pipenv==2020.8.13
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0